fix: export handling cleanup and declaration fixes
