+++ demo-angular/package.json
@@ -8,17 +8,17 @@
     "e2e-watch": "tsc -p e2e --watch"
   },
   "dependencies": {
-    "@angular/animations": "~10.1.0",
-    "@angular/common": "~10.1.0",
-    "@angular/compiler": "~10.1.0",
-    "@angular/core": "~10.1.0",
-    "@angular/forms": "~10.1.0",
-    "@angular/platform-browser": "~10.1.0",
-    "@angular/platform-browser-dynamic": "~10.1.0",
-    "@angular/router": "~10.1.0",
-    "@nativescript/angular": "^10.0.3",
-    "@nativescript/core": "^7.0.3",
-    "@nativescript/theme": "^2.3.3",
+    "@angular/animations": "~11.2.0",
+    "@angular/common": "~11.2.0",
+    "@angular/compiler": "~11.2.0",
+    "@angular/core": "~11.2.0",
+    "@angular/forms": "~11.2.0",
+    "@angular/platform-browser": "~11.2.0",
+    "@angular/platform-browser-dynamic": "~11.2.0",
+    "@angular/router": "~11.2.0",
+    "@nativescript/angular": "~11.8.0",
+    "@nativescript/core": "^8.0.0",
+    "@nativescript/theme": "^3.0.0",
     "@nativescript/webpack": "^3.0.4",
     "@nativescript-community/perms": "^2.1.1",
     "@nativescript-community/ui-mapbox": "file:../plugin",
@@ -26,12 +26,12 @@
     "reflect-metadata": "~0.1.12",
     "rxjs": "^6.4.0",
     "rxjs-compat": "^6.5.3",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "~10.1.0",
-    "@nativescript/android": "7.0.0",
-    "@ngtools/webpack": "~10.1.0",
+    "@angular/compiler-cli": "~11.2.0",
+    "@nativescript/android": "8.0.0",
+    "@ngtools/webpack": "~11.2.0",
     "@types/chai": "~4.1.7",
     "@types/mocha": "~5.2.5",
     "@types/node": "~10.12.18",
@@ -44,7 +44,7 @@
     "nativescript-dev-appium": "^6.1.2",
     "node-sass": "^4.7.1",
     "tslint": "~5.19.0",
-    "typescript": "~3.9.0"
+    "typescript": "~4.0.0"
   },
   "gitHead": "12284faf57b80a564bca5377e12c149c89d282c3",
   "readme": "NativeScript Application",
