+++ demo/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "@nativescript/core": "^7.0.13",
+    "@nativescript/core": "^8.0.0",
     "@nativescript-community/perms": "^2.1.1",
     "@nativescript/webpack": "3.0.8",
     "@nativescript-community/ui-mapbox": "file:../plugin"
@@ -8,10 +8,10 @@
   "devDependencies": {
     "@babel/core": "~7.12.3",
     "@babel/preset-env": "~7.12.1",
-    "@nativescript/android": "7.0.1",
-    "@nativescript/ios": "7.0.5",
-    "@nativescript/types": "7.0.4",
-    "typescript": "~3.9.7"
+    "@nativescript/android": "8.0.0",
+    "@nativescript/ios": "8.0.0",
+    "@nativescript/types": "8.0.0",
+    "typescript": "~4.0.0"
   },
   "main": "app.js"
 }
