+++ package.json
@@ -18,9 +18,10 @@
         "setup": "ts-patch install",
         "prepare": "npm run setup",
         "build.angular": "ng-packagr -p src/angular/package.json -c src/angular/tsconfig.json",
-        "clean": "rimraf ./plugin/**/*.d.ts ./plugin/**/*.js ./plugin/**/*.js.map plugin/node_modules plugin/package-lock.json plugin/README.md demo/node_modules demo-angular/node_modules",
+        "clean": "npx rimraf ./plugin/**/*.d.ts ./plugin/**/*.js ./plugin/**/*.js.map plugin/node_modules plugin/package-lock.json plugin/README.md demo/node_modules demo-angular/node_modules",
+        "reset": "npm run clean && yarn && npm run setup",
         "tsc": "cpy '**/*.d.ts' '../plugin' --parents --cwd=src && tsc -skipLibCheck -d",
-        "build": "cp README.md ./plugin && rm -f .tsbuildinfo && npm run tsc ",
+        "build": "cp README.md ./plugin && rm -f .tsbuildinfo && npm run tsc",
         "publish": "npm run setup && npm run build && lerna publish --create-release=github",
         "commitmsg": "commitlint -e $GIT_PARAMS",
         "demo.ios": "npm run build && cd ./demo && ns run ios",
@@ -66,21 +67,21 @@
     "homepage": "https://github.com/nativescript-community/ui-mapbox",
     "readmeFilename": "README.md",
     "devDependencies": {
-        "@angular/animations": "~10.1.5",
-        "@angular/common": "~10.1.5",
-        "@angular/compiler": "~10.1.5",
-        "@angular/compiler-cli": "~10.1.5",
-        "@angular/core": "~10.1.5",
-        "@angular/forms": "~10.1.5",
-        "@angular/platform-browser": "~10.1.5",
-        "@angular/platform-browser-dynamic": "11.0.0-rc.3",
-        "@angular/router": "~10.1.5",
+        "@angular/animations": "~11.2.0",
+        "@angular/common": "~11.2.0",
+        "@angular/compiler": "~11.2.0",
+        "@angular/compiler-cli": "~11.2.0",
+        "@angular/core": "~11.2.0",
+        "@angular/forms": "~11.2.0",
+        "@angular/platform-browser": "~11.2.0",
+        "@angular/platform-browser-dynamic": "~11.2.0",
+        "@angular/router": "~11.2.0",
         "@commitlint/cli": "^11.0.0",
         "@commitlint/config-conventional": "^11.0.0",
         "@nativescript-community/perms": "^2.1.1",
-        "@nativescript/angular": "10.1.7",
-        "@nativescript/core": "7.0.13",
-        "@nativescript/ios": "7.0.5",
+        "@nativescript/angular": "~11.8.0",
+        "@nativescript/core": "~8.0.0",
+        "@nativescript/ios": "8.0.0",
         "@nativescript/types-android": "~7.0.4",
         "@nativescript/types-ios": "~7.0.4",
         "@nativescript/webpack": "3.0.8",
@@ -94,7 +95,7 @@
         "eslint-plugin-prettier": "^3.1.4",
         "husky": "^4.3.0",
         "lerna": "^3.22.1",
-        "ng-packagr": "~10.1.2",
+        "ng-packagr": "~11.2.0",
         "prettier": "^2.1.2",
         "prompt": "^1.0.0",
         "react-nativescript": "^2.1.0",
@@ -103,7 +104,7 @@
         "svelte-native": "0.9.5",
         "ts-patch": "^1.3.0",
         "tslib": "2.0.3",
-        "typescript": "3.9.7",
+        "typescript": "~4.0.0",
         "zone.js": "0.11.3",
         "vue": "^2.6.12"
     },
