+++ plugin/platforms/ios/Podfile
@@ -1,3 +1,3 @@
-platform :ios, '10.0'
+platform :ios, '13.0'
 
 pod 'Mapbox-iOS-SDK'
