+++ src/filter/filter-parser.d.ts
@@ -1,0 +1,4 @@
+export declare class FilterParser {
+    static parseJson(json: any[]): any;
+    static toJson(filter: any): any[];
+}
