+++ src/layers/layer-factory.d.ts
@@ -1,5 +1,17 @@
-import { LayerCommon } from '..';
+import { LayerCommon } from '../common';
 
 declare class LayerFactory {
     static createLayer(style, source): Promise<LayerCommon>;
+}
+
+export declare class Layer implements LayerCommon {
+    id: string;
+    private instance;
+    constructor(instance: any);
+    visibility(): boolean;
+    show(): void;
+    hide(): void;
+    getNativeInstance(): any;
+    setFilter(filter: any[]): void;
+    getFilter(): any[];
 }
