+++ src/layers/parser/property-parser.d.ts
@@ -1,0 +1,3 @@
+export declare class PropertyParser {
+    static parsePropertiesForLayer(propertiesObject: any): any;
+}
